requirements.txt
django_extensions
django-import-export
celery
django-celery
django-crontab
mailcatcher
decouple